tweak sequencer webmaker
